// src/generated/types.ts
export interface HugoBuildOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** include content marked as draft */
  buildDrafts?: boolean;
  /** include expired content */
  buildExpired?: boolean;
  /** include content with publishdate in the future */
  buildFuture?: boolean;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** remove files from destination not found in static directories */
  cleanDestinationDir?: boolean;
  /** filesystem path to content directory */
  contentDir?: string;
  /** disable different kind of pages (home, RSS etc.) */
  disableKinds?: string[];
  /** add Git revision, date, author, and CODEOWNERS info to the pages */
  enableGitInfo?: boolean;
  /** copy all files when static is changed. */
  forceSyncStatic?: boolean;
  /** enable to run some cleanup tasks (remove unused cache files) after the build */
  gc?: boolean;
  /** ignore the configured file caches */
  ignoreCache?: boolean;
  /** filesystem path to layout directory */
  layoutDir?: string;
  /** minify any supported output format (HTML, XML etc.) */
  minify?: boolean;
  /** don't sync permission mode of files */
  noChmod?: boolean;
  /** don't sync modification time of files */
  noTimes?: boolean;
  /** panic on first WARNING log */
  panicOnWarning?: boolean;
  /** set this to a poll interval, e.g --poll 700ms, to use a poll based approach to watch for file system changes */
  poll?: string;
  /** print missing translations */
  printI18nWarnings?: boolean;
  /** print memory usage to screen at intervals */
  printMemoryUsage?: boolean;
  /** print warnings on duplicate target paths etc. */
  printPathWarnings?: boolean;
  /** print warnings on unused templates. */
  printUnusedTemplates?: boolean;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** display metrics about template executions */
  templateMetrics?: boolean;
  /** calculate some improvement hints when combined with --templateMetrics */
  templateMetricsHints?: boolean;
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
  /** write trace to file (not useful in general) */
  trace?: string;
  /** watch filesystem for changes and recreate as needed */
  watch?: boolean;
}

export interface HugoCompletionOptions extends HugoGlobalOptions {
}

export interface HugoCompletionBashOptions extends HugoGlobalOptions {
  /** disable completion descriptions */
  noDescriptions?: boolean;
}

export interface HugoCompletionFishOptions extends HugoGlobalOptions {
  /** disable completion descriptions */
  noDescriptions?: boolean;
}

export interface HugoCompletionPowershellOptions extends HugoGlobalOptions {
  /** disable completion descriptions */
  noDescriptions?: boolean;
}

export interface HugoCompletionZshOptions extends HugoGlobalOptions {
  /** disable completion descriptions */
  noDescriptions?: boolean;
}

export interface HugoConfigOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** preferred file format (toml, yaml or json) (default "toml") */
  format?: string;
  /** the language to display config for. Defaults to the first language defined. */
  lang?: string;
  /** include config options with zero values (e.g. false, 0, "") in the output */
  printZero?: boolean;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoConfigMountsOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoConvertOptions extends HugoGlobalOptions {
}

export interface HugoConvertToJSONOptions extends HugoGlobalOptions {
}

export interface HugoConvertToTOMLOptions extends HugoGlobalOptions {
}

export interface HugoConvertToYAMLOptions extends HugoGlobalOptions {
}

export interface HugoEnvOptions extends HugoGlobalOptions {
}

export interface HugoGenOptions extends HugoGlobalOptions {
}

export interface HugoGenDocOptions extends HugoGlobalOptions {
  /** the directory to write the doc. (default "/tmp/hugodoc/") */
  dir?: string;
}

export interface HugoGenManOptions extends HugoGlobalOptions {
  /** the directory to write the man pages. (default "man/") */
  dir?: string;
}

export interface HugoHelpOptions extends HugoGlobalOptions {
}

export interface HugoImportOptions extends HugoGlobalOptions {
}

export interface HugoImportJekyllOptions extends HugoGlobalOptions {
  /** allow import into non-empty target directory */
  force?: boolean;
}

export interface HugoListOptions extends HugoGlobalOptions {
}

export interface HugoListAllOptions extends HugoGlobalOptions {
}

export interface HugoListDraftsOptions extends HugoGlobalOptions {
}

export interface HugoListExpiredOptions extends HugoGlobalOptions {
}

export interface HugoListFutureOptions extends HugoGlobalOptions {
}

export interface HugoListPublishedOptions extends HugoGlobalOptions {
}

export interface HugoModOptions extends HugoGlobalOptions {
}

export interface HugoModCleanOptions extends HugoGlobalOptions {
  /** clean entire module cache */
  all?: boolean;
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** pattern matching module paths to clean (all if not set), e.g. "**hugo*" */
  pattern?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoModGetOptions extends HugoGlobalOptions {
}

export interface HugoModGraphOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** delete module cache for dependencies that fail verification */
  clean?: boolean;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoModInitOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoModNpmOptions extends HugoGlobalOptions {
}

export interface HugoModNpmPackOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoModTidyOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}

export interface HugoModVendorOptions extends HugoGlobalOptions {
  /** hostname (and path) to the root, e.g. https://spf13.com/ */
  baseURL?: string;
  /** filesystem path to cache directory */
  cacheDir?: string;
  /** filesystem path to content directory */
  contentDir?: string;
  /** named segments to render (configured in the segments config) */
  renderSegments?: string[];
  /** themes to use (located in /themes/THEMENAME/) */
  theme?: string[];
}
// ...
export interface HugoNewOptions extends HugoGlobalOptions {
}
// ...
export interface HugoNewProjectOptions extends HugoGlobalOptions {
  /** init inside non-empty directory */
  force?: boolean;
  /** preferred file format (toml, yaml or json) (default "toml") */
  format?: string;
}

export interface HugoNewThemeOptions extends HugoGlobalOptions {
  /** preferred file format (toml, yaml or json) (default "toml") */
  format?: string;
}
// ...
export interface HugoVersionOptions extends HugoGlobalOptions {
}

export type HugoCommand = "build" | "completion" | "completion bash" | "completion fish" | "completion powershell" | "completion zsh" | "config" | "config mounts" | "convert" | "convert toJSON" | "convert toTOML" | "convert toYAML" | "env" | "gen" | "gen doc" | "gen man" | "help" | "import" | "import jekyll" | "list" | "list all" | "list drafts" | "list expired" | "list future" | "list published" | "mod" | "mod clean" | "mod get" | "mod graph" | "mod init" | "mod npm" | "mod npm pack" | "mod tidy" | "mod vendor" | "mod verify" | "new" | "new content" | "new project" | "new theme" | "server" | "server trust" | "version";

export type HugoOptionsFor<C extends HugoCommand> =
  C extends "build" ? HugoBuildOptions :
  C extends "completion" ? HugoCompletionOptions :
  C extends "completion bash" ? HugoCompletionBashOptions :
  C extends "completion fish" ? HugoCompletionFishOptions :
  C extends "completion powershell" ? HugoCompletionPowershellOptions :
  C extends "completion zsh" ? HugoCompletionZshOptions :
  C extends "config" ? HugoConfigOptions :
  C extends "config mounts" ? HugoConfigMountsOptions :
  C extends "convert" ? HugoConvertOptions :
  C extends "convert toJSON" ? HugoConvertToJSONOptions :
  C extends "convert toTOML" ? HugoConvertToTOMLOptions :
  C extends "convert toYAML" ? HugoConvertToYAMLOptions :
  C extends "env" ? HugoEnvOptions :
  C extends "gen" ? HugoGenOptions :
  C extends "gen doc" ? HugoGenDocOptions :
  C extends "gen man" ? HugoGenManOptions :
  C extends "help" ? HugoHelpOptions :
  C extends "import" ? HugoImportOptions :
  C extends "import jekyll" ? HugoImportJekyllOptions :
  C extends "list" ? HugoListOptions :
  C extends "list all" ? HugoListAllOptions :
  C extends "list drafts" ? HugoListDraftsOptions :
  C extends "list expired" ? HugoListExpiredOptions :
  C extends "list future" ? HugoListFutureOptions :
  C extends "list published" ? HugoListPublishedOptions :
  C extends "mod" ? HugoModOptions :
  C extends "mod clean" ? HugoModCleanOptions :
  C extends "mod get" ? HugoModGetOptions :
  C extends "mod graph" ? HugoModGraphOptions :
  C extends "mod init" ? HugoModInitOptions :
  C extends "mod npm" ? HugoModNpmOptions :
  C extends "mod npm pack" ? HugoModNpmPackOptions :
  C extends "mod tidy" ? HugoModTidyOptions :
  C extends "mod vendor" ? HugoModVendorOptions :
  C extends "mod verify" ? HugoModVerifyOptions :
  C extends "new" ? HugoNewOptions :
  C extends "new content" ? HugoNewContentOptions :
  C extends "new project" ? HugoNewProjectOptions :
  C extends "new theme" ? HugoNewThemeOptions :
  C extends "server" ? HugoServerOptions :
  C extends "server trust" ? HugoServerTrustOptions :
  C extends "version" ? HugoVersionOptions :
  never;